+++ src/index.js
@@ -1,3 +1,4 @@
+export * from './dom/index.js'
 export * from './math/index.js'
 export * from './string/index.js'
 export * from './url/email.js'