+++ src/math/index.js
@@ -13,8 +13,11 @@
  * @param {number} val The number you want to maintain inside boundaries.
  * @param {number} min The lowest permitted value.
  * @param {number} max The highest permitted value.
+ * @returns {number}
  */
-export const clamp = (val, min, max) => Math.max(min, Math.min(max, val))
+export const clamp = (val, min, max) => min > max
+  ? clamp(val, max, min)
+  : Math.max(min, Math.min(max, val))
 
 /**
  * Round a number to the (optionally) provided precision.